configfile: Add support for XDG_CONFIG_HOME
This allows bugz config to be stored in the XDG_CONFIG_HOME directory
(which defaults to ~/.config). Per the XDG basedir-spec
(https://standards.freedesktop.org/basedir-spec/basedir-spec-0.6.html):

"""
$XDG_CONFIG_HOME defines the base directory relative to which user
specific configuration files should be stored. If $XDG_CONFIG_HOME is
either not set or empty, a default equal to $HOME/.config should be
used.
"""

This patch checks for the existence of each file in order, loading only
one, starting with XDG_CONFIG_HOME, moving to the fallback, and finally
to the *nix standard ~/.bugzrc, and handles the case of none of these
existing.